fix unload
